[test/fix] Fix signed-unsigned compare
- Fix build fail issue when build with capi

+++ tests/tizen_capi/unittest_tizen_capi_latency.cc
@@ -167,7 +167,7 @@ class nnstreamer_capi_singleshot_latency : public ::testing::Test
     if (fd >= 0) {
       resetDataFile ();
       data_read = read (fd, data, data_size);
-      EXPECT_EQ (data_read, data_size);
+      EXPECT_EQ ((size_t) data_read, data_size);
     }
 
     /** Benchmark the invoke duration */
